organize code of part text2image
