+++ examples/stable_diffusion_v2/ldm/models/diffusion/plms.py
@@ -15,7 +15,6 @@
 import logging
 
 from ldm.modules.diffusionmodules.util import make_ddim_sampling_parameters, make_ddim_timesteps, noise_like
-from ldm.util import is_old_ms_version
 from tqdm import tqdm
 
 import mindspore as ms
@@ -255,10 +254,7 @@ class PLMSSampler:
                     c_in = ops.concat((unconditional_conditioning, c), axis=0)
                     ldm_output = self.model.apply_model(x_in, t_in, c_crossattn=c_in)
 
-                if is_old_ms_version():
-                    e_t_uncond, e_t = ops.split(ldm_output, axis=0, output_num=2)
-                else:
-                    e_t_uncond, e_t = ops.split(ldm_output, split_size_or_sections=ldm_output.shape[0] // 2, axis=0)
+                e_t_uncond, e_t = ops.split(ldm_output, split_size_or_sections=ldm_output.shape[0] // 2, axis=0)
 
                 e_t = e_t_uncond + unconditional_guidance_scale * (e_t - e_t_uncond)
