+++ examples/stable_diffusion_v2/ldm/modules/diffusionmodules/openaimodel.py
@@ -24,7 +24,6 @@ from ldm.modules.diffusionmodules.util import (
     timestep_embedding,
     zero_module,
 )
-from ldm.util import is_old_ms_version
 
 import mindspore as ms
 import mindspore.nn as nn
@@ -162,11 +161,7 @@ class ResBlock(nn.Cell):
 
         self.out_layers_norm = normalization(self.out_channels)
         self.out_layers_silu = nn.SiLU().to_float(self.dtype)
-
-        if is_old_ms_version():
-            self.out_layers_drop = nn.Dropout(keep_prob=self.dropout)
-        else:
-            self.out_layers_drop = nn.Dropout(p=1.0 - self.dropout)
+        self.out_layers_drop = nn.Dropout(p=1.0 - self.dropout)
 
         self.out_layers_conv = zero_module(
             conv_nd(dims, self.out_channels, self.out_channels, 3, padding=1, has_bias=True, pad_mode="pad").to_float(
